+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added: `unit-whitelist` rule.
 * Added: `property-unit-blacklist` rule.
 * Fixed: bug when loading plugins from an extended config
+* Fixed: bug causing `function-comma-*` whitespace rules to improperly judge whether to enforce single- or multi-line options.
 
 # 2.0.0